functions/shared/requirements.txt
cloud-sql-python-connector[pymysql]==1.4.3
SQLAlchemy==2.0.23
PyMySQL==1.1.0
google-cloud-secret-manager==2.16.4
google-cloud-datastore==2.19.0

